Extract caching feature flag into its own class.
This also shrinks the abstract class to just its abstract interface.

Closes #95
